+++ packages/component/jest.config.js
@@ -1,8 +1,0 @@
-module.exports = {
-  collectCoverage: true,
-  coverageReporters: [
-    'lcov',
-    'text-summary'
-  ],
-  testURL: 'http://localhost/'
-};