+++ packages/integration-test/createDeferred.js
@@ -1,0 +1,15 @@
+/* eslint import/no-commonjs: "off" */
+module.exports = function createDeferred() {
+  let reject, resolve;
+
+  const promise = new Promise((promiseResolve, promiseReject) => {
+    reject = promiseReject;
+    resolve = promiseResolve;
+  });
+
+  if (!reject || !resolve) {
+    throw new Error('Promise is not a ES-compliant and do not run exector immediately');
+  }
+
+  return { promise, reject, resolve };
+};