+++ packages/react-say/src/index.mjs
@@ -1,0 +1,9 @@
+import Composer from './Composer.jsx';
+import Context from './Context.mjs';
+import Say from './Say.jsx';
+import SayButton from './SayButton.jsx';
+import SayUtterance from './SayUtterance.jsx';
+
+export default Say;
+
+export { Composer, Context, SayButton, SayUtterance };