+++ packages/react-say/src/index.test.ts
@@ -1,0 +1,3 @@
+import { test } from 'node:test';
+
+test('Setup', () => {});