+++ packages/react-say/src/retry.mjs
@@ -1,0 +1,19 @@
+import sleep from './sleep.mjs';
+
+export default async function retry(fn, numRetries, interval) {
+  let lastError;
+
+  for (let times = 0; times < numRetries; times++) {
+    if (times) {
+      await sleep(interval);
+    }
+
+    try {
+      return await fn();
+    } catch (error) {
+      lastError = error;
+    }
+  }
+
+  throw lastError;
+}